+++ src/syscalls/interfaces/mod.rs
@@ -1,3 +1,1 @@
-pub use self::uhyve::*;
-
 pub(crate) mod uhyve;
